+++ src/pie_core/hf_hub_mixin.py
@@ -5,7 +5,6 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Type, TypeVar, Union
 
 import requests
-from huggingface_hub.constants import CONFIG_NAME, PYTORCH_WEIGHTS_NAME
 from huggingface_hub.file_download import hf_hub_download
 from huggingface_hub.hf_api import HfApi
 from huggingface_hub.utils import SoftTemporaryDirectory, validate_hf_hub_args
@@ -14,10 +13,6 @@ from pie_core.utils.dictionary import TNestedBoolDict, dict_update_nested
 
 logger = logging.getLogger(__name__)
 
-MODEL_CONFIG_NAME = CONFIG_NAME
-TASKMODULE_CONFIG_NAME = "taskmodule_config.json"
-MODEL_CONFIG_TYPE_KEY = "model_type"
-TASKMODULE_CONFIG_TYPE_KEY = "taskmodule_type"
 
 # Generic variable that is either PieBaseHFHubMixin or a subclass thereof
 T = TypeVar("T", bound="PieBaseHFHubMixin")
@@ -31,8 +26,8 @@ class PieBaseHFHubMixin:
     of mixin integration with the Hub. Check out our [integration guide](../guides/integrations) for more instructions.
     """
 
-    config_name = MODEL_CONFIG_NAME
-    config_type_key = MODEL_CONFIG_TYPE_KEY
+    config_name = "not_implemented.json"
+    config_type_key = "not_implemented"
 
     def __init__(self, *args, is_from_pretrained: bool = False, **kwargs):
         super().__init__(*args, **kwargs)
@@ -101,7 +96,7 @@ class PieBaseHFHubMixin:
             save_directory (`str` or `Path`):
                 Path to directory in which the model weights and configuration will be saved.
         """
-        raise NotImplementedError
+        return None
 
     @classmethod
     def retrieve_config_file(
@@ -237,7 +232,8 @@ class PieBaseHFHubMixin:
         resume_download: bool,
         local_files_only: bool,
         token: Optional[Union[str, bool]],
-        **model_kwargs,
+        config: Optional[dict] = None,
+        **kwargs,
     ) -> T:
         """Overwrite this method in subclass to define how to load your model from pretrained.
 
@@ -272,7 +268,9 @@ class PieBaseHFHubMixin:
             model_kwargs:
                 Additional keyword arguments passed along to the [`~ModelHubMixin._from_pretrained`] method.
         """
-        raise NotImplementedError
+        module = cls.from_config(config=config or {}, **kwargs)
+
+        return module
 
     @validate_hf_hub_args
     def push_to_hub(
@@ -375,205 +373,3 @@ class PieBaseHFHubMixin:
         config = config.copy()
         dict_update_nested(config, kwargs, override=config_override)
         return cls(**config)
-
-
-TModel = TypeVar("TModel", bound="PieModelHFHubMixin")
-
-
-class PieModelHFHubMixin(PieBaseHFHubMixin):
-    config_name = MODEL_CONFIG_NAME
-    config_type_key = MODEL_CONFIG_TYPE_KEY
-    weights_file_name = PYTORCH_WEIGHTS_NAME
-    """Implementation of [`PieBaseHFHubMixin`] to provide model Hub upload/download capabilities to
-    models.
-
-    Example for a Pytorch model:
-
-    ```python
-    >>> import torch
-    >>> import torch.nn as nn
-    >>> from pie_core import PieModelHFHubMixin
-
-
-    >>> class MyPytorchModel(nn.Module, PieModelHFHubMixin):
-    ...     def __init__(self):
-    ...         super().__init__()
-    ...         self.param = nn.Parameter(torch.rand(3, 4))
-    ...         self.linear = nn.Linear(4, 5)
-
-    ...     def forward(self, x):
-    ...         return self.linear(x + self.param)
-    ...
-    ...     def save_model_file(self, model_file: str) -> None:
-    ...         torch.save(self.state_dict(), model_file)
-    ...
-    ...     def load_model_file(
-    ...         self, model_file: str, map_location: str = "cpu", strict: bool = False
-    ...     ) -> None:
-    ...         state_dict = torch.load(model_file, map_location=torch.device(map_location))
-    ...         self.load_state_dict(state_dict, strict=strict)
-
-    >>> model = MyModel()
-
-    # Save model weights to local directory
-    >>> model.save_pretrained("my-awesome-model")
-
-    # Push model weights to the Hub
-    >>> model.push_to_hub("my-awesome-model")
-
-    # Download and initialize weights from the Hub
-    >>> model = MyModel.from_pretrained("username/my-awesome-model")
-    ```
-    """
-
-    def save_model_file(self, model_file: str) -> None:
-        """Save weights from a Pytorch model to a local directory."""
-        raise NotImplementedError
-
-    def load_model_file(
-        self, model_file: str, map_location: str = "cpu", strict: bool = False
-    ) -> None:
-        """Load weights from a Pytorch model file."""
-        raise NotImplementedError
-
-    def _save_pretrained(self, save_directory: Path) -> None:
-        """Save weights from a Pytorch model to a local directory."""
-        self.save_model_file(str(save_directory / self.weights_file_name))
-
-    @classmethod
-    def retrieve_model_file(
-        cls,
-        model_id: str,
-        revision: Optional[str] = None,
-        cache_dir: Optional[Union[str, Path]] = None,
-        force_download: bool = False,
-        proxies: Optional[Dict] = None,
-        resume_download: bool = False,
-        local_files_only: bool = False,
-        token: Optional[Union[str, bool]] = None,
-        **remaining_kwargs,
-    ) -> Tuple[str, Dict[str, Any]]:
-        """Retrieve the model file from the Huggingface Hub or local directory."""
-        if os.path.isdir(model_id):
-            logger.info("Loading weights from local directory")
-            model_file = os.path.join(model_id, cls.weights_file_name)
-        else:
-            model_file = hf_hub_download(
-                repo_id=model_id,
-                filename=cls.weights_file_name,
-                revision=revision,
-                cache_dir=cache_dir,
-                force_download=force_download,
-                proxies=proxies,
-                resume_download=resume_download,
-                token=token,
-                local_files_only=local_files_only,
-            )
-
-        return model_file, remaining_kwargs
-
-    @classmethod
-    def _from_pretrained(
-        cls: Type[TModel],
-        *,
-        model_id: str,
-        revision: Optional[str],
-        cache_dir: Optional[Union[str, Path]],
-        force_download: bool,
-        proxies: Optional[Dict],
-        resume_download: bool,
-        local_files_only: bool,
-        token: Union[str, bool, None],
-        map_location: str = "cpu",
-        strict: bool = False,
-        config: Optional[dict] = None,
-        **kwargs,
-    ) -> TModel:
-
-        model_file, remaining_kwargs = cls.retrieve_model_file(
-            model_id=model_id,
-            revision=revision,
-            cache_dir=cache_dir,
-            force_download=force_download,
-            proxies=proxies,
-            resume_download=resume_download,
-            local_files_only=local_files_only,
-            token=token,
-            **kwargs,
-        )
-        model = cls.from_config(config=config or {}, **remaining_kwargs)
-
-        # TODO: map_location and strict are quite specific to PyTorch.
-        #  How to handle this in a more generic way?
-        model.load_model_file(model_file, map_location=map_location, strict=strict)
-
-        return model
-
-
-TTaskModule = TypeVar("TTaskModule", bound="PieTaskModuleHFHubMixin")
-
-
-class PieTaskModuleHFHubMixin(PieBaseHFHubMixin):
-    config_name = TASKMODULE_CONFIG_NAME
-    config_type_key = TASKMODULE_CONFIG_TYPE_KEY
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def _save_pretrained(self, save_directory) -> None:
-        return None
-
-    @classmethod
-    def _from_pretrained(
-        cls: Type[TTaskModule],
-        *,
-        model_id: str,
-        revision: Optional[str],
-        cache_dir: Optional[Union[str, Path]],
-        force_download: bool,
-        proxies: Optional[Dict],
-        resume_download: bool,
-        local_files_only: bool,
-        token: Union[str, bool, None],
-        map_location: str = "cpu",
-        strict: bool = False,
-        config: Optional[dict] = None,
-        **kwargs,
-    ) -> TTaskModule:
-
-        taskmodule = cls.from_config(config=config or {}, **kwargs)
-
-        return taskmodule
-
-
-TPipeline = TypeVar("TPipeline", bound="AnnotationPipelineHFHubMixin")
-
-
-class AnnotationPipelineHFHubMixin(PieBaseHFHubMixin):
-    config_name = "pipeline_config.json"
-    config_type_key = "pipeline_type"
-
-    def _save_pretrained(self, save_directory) -> None:
-        return None
-
-    @classmethod
-    def _from_pretrained(
-        cls: Type[TPipeline],
-        *,
-        model_id: str,
-        revision: Optional[str],
-        cache_dir: Optional[Union[str, Path]],
-        force_download: bool,
-        proxies: Optional[Dict],
-        resume_download: bool,
-        local_files_only: bool,
-        token: Union[str, bool, None],
-        map_location: str = "cpu",
-        strict: bool = False,
-        config: Optional[dict] = None,
-        **kwargs,
-    ) -> TPipeline:
-
-        pipeline = cls.from_config(config=config or {}, **kwargs)
-
-        return pipeline
